refactor(flag): 举报流程迁 v2 服务端动态类型 + FlagReason 迁入 feeds 框架
举报第一步的违规类型不再写死 4 个本地字符串（服务端早已不认、一律 403），改走
/v1/illust/report/topic-list 动态下发，提交走 /v2/illust/report 的 topic_id +
description。顺带修掉一个从未生效过的老 bug：flagObjectId 历史上用 Int extra 存/读，
跟 Illust.id(Long) 类型不匹配，Bundle 静默返回 0，提交的 illust_id 恒为 0——各入口
（FragmentIllust / ArtworkV3Fragment / TemplateActivity）统一改 Long。

FlagReasonFragment 从「在 Fragment 里 launch 协程发请求 + 手 inflate 行」迁入 feeds
框架：请求搬进零捕获 FeedSource（数据归 VM，旋转存活），刷新/空态/错误重试交给框架。
hero 圆底做成 fullSpan header 条目，分段行圆角背景在无分页整表已知时于数据源按位置预算
好塞进 FlagTopicFeedItem。横向留白走 ItemDecoration（setUpToolbar 的 inset 监听会抹掉
列表 padding，跟 PrimeTags 等 feeds 页一致），刻意不加行间纵向 inset 以免拆散分段圆角。

退休旧 CommonAdapter 体系的 FlagReason/FlagReasonHolder/cell_flag_reason.xml。

+++ app/src/main/java/ceui/lisa/fragments/FragmentIllust.kt
@@ -321,7 +321,10 @@ class FragmentIllust : SwipeFragment<FragmentIllustBinding>() {
                 R.id.action_flag_illust -> {
                     val intent = Intent(mContext, TemplateActivity::class.java)
                     intent.putExtra(TemplateActivity.EXTRA_FRAGMENT, "举报插画")
-                    intent.putExtra(FlagDescFragment.FlagObjectIdKey, illust.id)
+                    // illust.id 是 IllustsBean 的 Int 字段；TemplateActivity 读这个 extra 走
+                    // getLongExtra（配合 Illust.id: Long 的新版详情页入口），这里必须显式转 Long，
+                    // 否则又是一次 Int/Long extra 类型不匹配，读回来静默变 0。
+                    intent.putExtra(FlagDescFragment.FlagObjectIdKey, illust.id.toLong())
                     intent.putExtra(FlagDescFragment.FlagObjectTypeKey, ObjectSpec.POST)
                     startActivity(intent)
                     true
